ファイル、ディレクトリ構造の見直し
