+++ src/constants/sections.ts
@@ -1,0 +1,3 @@
+﻿import type { Section } from '../types/portfolio';
+
+export const SECTIONS: Section[] = ['hero', 'about', 'projects', 'contact'];
