+++ src/data/translations.ts
@@ -1,0 +1,34 @@
+﻿import type { Translations } from '../types/portfolio';
+
+export const translations: Translations = {
+  en: {
+    role: 'University Student / Web & AI Emginner',
+    explore: 'Explore',
+    details: 'View Details',
+    close: 'Close',
+    contact: 'Contact Me',
+    about: {
+      title: 'About Me',
+      desc: 'こんにちは！私は関西の大学に通う28卒の学生です。趣味はサウナ、バスケットボール、カラオケ、スノーボードです。大学ではWebアプリ開発に興味を持ち、特にReactを中心に学んできました。他にもLaravelやDjangoを扱った経験があり、バックエンド開発にも関心があります。優れたアプリケーションは見た目だけでなく、スムーズな操作性と実用性を兼ね備えているべきだと考えています。研究室では知能情報研究室に所属しており、主にAIのニューラルネットワークについて学んでいます。最近はTransformerに関する論文を読み、その理解を深めています。技術の進化に追いつきながら、より多くの分野に触れていくことを大切にしています。将来的には、AIの分野に積極的に関わっていきたいと考えています。新しい技術や挑戦に対して常にオープンな姿勢で取り組んでいますので、興味を持っていただけたらぜひお声がけください！',
+    },
+    projects: {
+      title: 'Selected Projects',
+      desc: 'Selected Projects',
+    }
+  },
+  ja: {
+    role: 'University Student / Web & AI Emginner',
+    explore: '探索する',
+    details: '詳細を見る',
+    close: '閉じる',
+    contact: 'コンタクト',
+    about: {
+      title: 'About Me',
+      desc: 'こんにちは！私は関西の大学に通う28卒の学生です。趣味はサウナ、バスケットボール、カラオケ、スノーボードです。大学ではWebアプリ開発に興味を持ち、特にReactを中心に学んできました。他にもLaravelやDjangoを扱った経験があり、バックエンド開発にも関心があります。優れたアプリケーションは見た目だけでなく、スムーズな操作性と実用性を兼ね備えているべきだと考えています。研究室では知能情報研究室に所属しており、主にAIのニューラルネットワークについて学んでいます。最近はTransformerに関する論文を読み、その理解を深めています。技術の進化に追いつきながら、より多くの分野に触れていくことを大切にしています。将来的には、AIの分野に積極的に関わっていきたいと考えています。新しい技術や挑戦に対して常にオープンな姿勢で取り組んでいますので、興味を持っていただけたらぜひお声がけください！',
+    },
+    projects: {
+      title: 'Selected Projects',
+      desc: 'Selected Projects',
+    }
+  }
+};
